+++ src/main/java/team05/integrated_feed_backend/module/post/service/PostService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import team05.integrated_feed_backend.exception.code.StatusCode;
+import team05.integrated_feed_backend.common.code.StatusCode;
 import team05.integrated_feed_backend.exception.custom.DataNotFoundException;
 import team05.integrated_feed_backend.module.post.entity.Post;
 import team05.integrated_feed_backend.module.post.event.publisher.PostEventPublisher;
